2020/202001_refactoring_workshop/tennis_test.py
import pytest


from tennis import TennisGame


@pytest.mark.parametrize('p1_points,p2_points,score', [
    (0, 0, "Love-All"),
    (1, 1, "Fifteen-All"),
    (2, 2, "Thirty-All"),
    (3, 3, "Forty-All"),
    (4, 4, "Deuce"),

    (1, 0, "Fifteen-Love"),
    (0, 1, "Love-Fifteen"),
    (2, 0, "Thirty-Love"),
    (0, 2, "Love-Thirty"),
    (3, 0, "Forty-Love"),
    (0, 3, "Love-Forty"),
    (4, 0, "Win for player1"),
    (0, 4, "Win for player2"),

    (2, 1, "Thirty-Fifteen"),
    (1, 2, "Fifteen-Thirty"),
    (3, 1, "Forty-Fifteen"),
    (1, 3, "Fifteen-Forty"),
    (4, 1, "Win for player1"),
    (1, 4, "Win for player2"),

    (3, 2, "Forty-Thirty"),
    (2, 3, "Thirty-Forty"),
    (4, 2, "Win for player1"),
    (2, 4, "Win for player2"),

    (4, 3, "Advantage player1"),
    (3, 4, "Advantage player2"),
    (5, 4, "Advantage player1"),
    (4, 5, "Advantage player2"),
    (15, 14, "Advantage player1"),
    (14, 15, "Advantage player2"),

    (6, 4, "Win for player1"),
    (4, 6, "Win for player2"),
    (16, 14, "Win for player1"),
    (14, 16, "Win for player2"),
])
def test_get_score(p1_points, p2_points, score, p1_name='player1', p2_name='player2'):
    game = TennisGame(p1_name, p2_name)
    for i in range(p1_points):
        game.won_point(p1_name)
    for i in range(p2_points):
        game.won_point(p2_name)
    assert score == game.score()


def test_player_names():
    game = TennisGame('One', 'Two')
    for i in range(4):
        game.won_point('One')
    assert "Win for One" == game.score()

